+++ action.yaml
@@ -6,4 +6,4 @@ branding:
   color: black
 runs:
   using: node20
-  main: dist/index.mjs
+  main: dist/index.js
