+++ src/index.ts
@@ -3,7 +3,7 @@ import core from "@actions/core";
 import { hashFile } from "hasha";
 import fs from "fs";
 import os from "os";
-import yarn from "./yarn.mjs";
+import yarn from "./yarn.js";
 
 async function main(): Promise<void> {
   await core.group("Enabling Yarn", async () => {
